+++ CHANGELOG.md
@@ -14,6 +14,10 @@ and uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ### Changed
 * metadata tile script name from `generate_frame_tile.py` to `generate_metadata_tile.py`
 * `generate_frame_tile.py` CLI to take upper left corner of bounding box, instead of full bounding box, as input
+* Frame ordering strategy so that west most relative orbits are on top for ascending data, and on the bottom for descending data
+
+### Fixed
+* Removal of empty metadata tile is now conditional on the tile being all nodata, not the absence of granules
 
 ## [0.1.2]
 ### Added